refactor: ApiResponse 클래스 패키지 위치 변경
- ApiResponse 클래스를 common 패키지로 이동
- 패키지 경로 변경에 따른 import 구문 수정
- 기존 코드에서 ApiResponse 경로 의존성 업데이트
